LayoutTests/touchadjustment/disabled-formelements.html

<html>
<head>
<title>Touch Adjustment : Disabled form elements - bug 92093</title>
<style type="text/css">
    #sandbox {
        position: absolute;
        left: 0px;
        top: 0px;
        width: 600px;
        height: 600px;
    }
</style>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>

<body id="body">

<div id=sandbox>
    <form id="form">
        <label for="input1" id="label1">Label</label> <input type="text" id="input1" value="Disabled input" disabled></input><br>
        <label for="input2" id="label2">Label</label> <input type="text" id="input2" value="Enabled input"></input>
    </form>
</div>

<p id='description'></p>
<div id='console'></div>

<script>
    var adjustedNode;
    function testRoundTouch(x, y, radius)
    {
        var x = x - radius;
        var y = y - radius;
        var width = radius * 2;
        var height = radius * 2;
        var adjustedNode = internals.touchNodeAdjustedToBestClickableNode(x, y, width, height, document);
        if (adjustedNode.nodeType == 3) // TEXT node
            adjustedNode = adjustedNode.parentNode;
        return adjustedNode;
    }

    function testDirectTouches()
    {
        // Quick sanity check.
        debug('Test we can hit the form elements directly.');

//         adjustedNode = testRoundTouch(16, 8, 5);
//         shouldBeEqualToString('adjustedNode.id', 'label1');

        adjustedNode = testRoundTouch(16, 38, 5);
        shouldBeEqualToString('adjustedNode.id', 'label2');

//         adjustedNode = testRoundTouch(60, 8, 5);
//         shouldBeEqualToString('adjustedNode.id', 'input1');

        adjustedNode = testRoundTouch(60, 38, 5);
        shouldBeEqualToString('adjustedNode.id', 'input2');

    }

    function testAdjustedTouches()
    {
        debug('Test indirect touches are adjusted to the enabled form elements.');
        adjustedNode = testRoundTouch(16, 25, 10);
        shouldBeEqualToString('adjustedNode.id', 'label2');

        adjustedNode = testRoundTouch(16, 20, 15);
        shouldBeEqualToString('adjustedNode.id', 'label2');

        adjustedNode = testRoundTouch(16, 16, 20);
        shouldBeEqualToString('adjustedNode.id', 'label2');

        adjustedNode = testRoundTouch(60, 25, 10);
        shouldBeEqualToString('adjustedNode.id', 'input2');

        adjustedNode = testRoundTouch(60, 20, 15);
        shouldBeEqualToString('adjustedNode.id', 'input2');

        adjustedNode = testRoundTouch(60, 16, 20);
        shouldBeEqualToString('adjustedNode.id', 'input2');
    }

    function runTests()
    {
        if (window.testRunner && window.internals && internals.touchNodeAdjustedToBestClickableNode) {
            description(document.title);
            testRunner.dumpAsText();
            testDirectTouches();
            testAdjustedTouches();
        }
    }

    runTests();
</script>

<script src="../fast/js/resources/js-test-post.js"></script>

</body>
</html>
