LayoutTests/touchadjustment/disabled-formelements-expected.txt

Label 
Label 
Touch Adjustment : Disabled form elements - bug 92093

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test we can hit the form elements directly.
PASS adjustedNode.id is "label2"
PASS adjustedNode.id is "input2"
Test indirect touches are adjusted to the enabled form elements.
PASS adjustedNode.id is "label2"
PASS adjustedNode.id is "label2"
PASS adjustedNode.id is "label2"
PASS adjustedNode.id is "input2"
PASS adjustedNode.id is "input2"
PASS adjustedNode.id is "input2"
PASS successfullyParsed is true

TEST COMPLETE
